[REFACTOR] Move menu components into folder
