+++ src/client/react/user/components/menus/PostRaceMenu.js
@@ -3,15 +3,15 @@ import autobind from 'core-decorators/es/autobind';
 import { withRouter } from 'react-router-dom';
 import { Position, Spinner, Menu, MenuItem, MenuDivider, Intent } from '@blueprintjs/core';
 import { connect } from 'react-redux';
-import { logout } from '../../../actions/authActions';
+import { logout } from '../../../../actions/authActions';
 import axios from 'axios';
-import API from '../../../API';
-import MenuButton from '../../../../../lib/react/components/MenuButton';
-import LoginRefresher from '../../components/LoginRefresher';
-import LogoutFunction from '../../components/LogoutFunction';
-import NotificationToaster from '../../components/NotificationToaster';
+import API from '../../../../API';
+import MenuButton from '../../../../../../lib/react/components/MenuButton';
+import LoginRefresher from '../../../components/LoginRefresher';
+import LogoutFunction from '../../../components/LogoutFunction';
+import NotificationToaster from '../../../components/NotificationToaster';
 
-import '../scss/components/_help-menu.scss';
+import '../../scss/components/_help-menu.scss';
 
 
 const mapStateToProps = (state, ownProps) => {
